requirements.txt
# Core ML
tensorflow==2.18.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2
Pillow==10.2.0

# MLOps
mlflow==2.9.2
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Utilities
hydra-core==1.3.2
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0

# Visualization
gradio==4.15.0
matplotlib==3.8.2
seaborn==0.13.1